onlineshopping/requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
Django==2.1.3
django-celery-beat==1.2.0
django-timezone-field==3.0
djangorestframework==3.9.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kombu==4.2.1
lxml==4.2.5
mysqlclient==1.3.13
parsel==1.5.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-crontab==2.3.5
python-dateutil==2.7.5
python-scrapyd-api==2.1.2
pytz==2018.7
queuelib==1.5.0
redis==2.10.5
requests==2.20.0
Scrapy==1.5.1
scrapyd==1.2.0
service-identity==17.0.0
six==1.11.0
Twisted==18.9.0
urllib3==1.24
vine==1.1.4
w3lib==1.19.0
zope.interface==4.6.0
djangorestframework-filters==0.10.1
django-filter==2.0.0
